requirements.txt
django==2.1.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-debug-toolbar==1.11
django-filter==2.1.0
django-taggit==0.23.0
Pillow==5.3.0
pytz==2018.7
reportlab==3.5.12
sorl-thumbnail==12.5.0
sqlparse==0.2.4